refactor the certapi helpers to separate file
